Iframe: calc compat styles once per page load
